# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": "Microservice Posts",
    "description": "Microservice from the \"If community\" project for managing posts and their categories and interactions (upvotes)"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "api-token",
      "description": "An API token is a unique access key that grants permission to interact with the API, ensuring security and control over the operations performed."
    }
  },
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "adding new user to microservice",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "name": {
                  "type": "string",
                  "example": "joão da silva"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 5
                },
                "name": {
                  "type": "string",
                  "example": "joão da silva"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-19T11:11:43.799Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Um usuário com esse id já está cadastrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Conflict"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Users"
        ],
        "description": "lists all users registered in the system",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Number informing which page the search should return\n (pattern 1)",
            "type": "number"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number informing the number of elements per page\n (pattern 10)",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "totalPages": {
                  "type": "number",
                  "example": 1
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 2
                      },
                      "name": {
                        "type": "string",
                        "example": "maria da silva"
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-05-19T11:11:43.799Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "null"
                      },
                      "deletedAt": {
                        "type": "string",
                        "example": "null"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "search for a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "The user ID that was registered in the system"
          }
        ],
        "responses": {
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Um usuário com esse id não está cadastrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "description": "update user specified by ID",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "The user ID that was registered in the system"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "joão da silva pereira"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 5
                },
                "name": {
                  "type": "string",
                  "example": "joão da silva pereira"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-19T11:11:43.799Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Um usuário com esse id não está cadastrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "removes a user from the specified id",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "The user ID that was registered in the system"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "delete": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Usuário não encontrado para exclusão"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  }
}